pyproject.toml
[project]
name = "weather-collector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.210.0",
    "boto3>=1.38.27",
    "constructs>=10.4.2",
    "pandas>=2.2.3",
    "playwright>=1.54.0",
    "pyarrow>=19.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.8",
]
